requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
PyJWT[crypto]==2.10.1
bcrypt==4.2.1

# HTTP client
httpx==0.28.1

# Scheduler
apscheduler==3.10.4

# Settings
pydantic-settings==2.7.0
pydantic[email]==2.10.4

# Diff
dictdiffer==0.9.0

# Rate limiting
slowapi==0.1.9

# Utilities
python-multipart==0.0.20
